3619: Rename EntityAttributeEditor files
