+++ common/src/View/EntityPropertyEditor.h
